[www] Refactor `InformationCard` into common
Preparing to make it a sticky sidebar
