+++ packages/www/components/FirstPage/index.module.css
@@ -6,32 +6,3 @@
   width: 100vw;
   min-height: 100vh;
 }
-
-.InfoCard {
-  margin: 8px;
-}
-
-@media screen and (max-width: 720px) {
-  .InfoCard {
-    width: 90vw;
-  }
-}
-
-@media screen and (min-width: 720px) {
-  .InfoCard {
-    width: calc(100% - 500px);
-    margin-left: 50px;
-    max-width: 300px;
-  }
-}
-
-.IconLine {
-  display: flex;
-  align-items: center;
-  align-content: center;
-  line-height: 24px;
-}
-
-.IconLineText {
-  margin-left: 8px;
-}
